Action the round's ten findings across docs and tests
Name the three XDG labels in an XdgDecisions struct so a transposed
field is a compile error rather than mislabelled telemetry, and state
the unused_self suppression on the non-Unix stub instead of the dead
`_ = self;` assignment. Give every public env-source function an
outcome-focused doctest, expand RPITIT on first use, say ProcessEnv
overrides home_fallback by default rather than exclusively, scope the
RFC's no-lock claim to the policy snapshot test, and list each metrics
counter's labels separately. Pin XDG_CONFIG_DIRS to a bare base in the
compose_layers isolation builder so a host /etc/xdg/demo file cannot
leak into the fixtures, split the metrics-facade suite into its own
integration test to bring discovery_telemetry.rs under the 400-line
ceiling, and return the segment regex as the strategy it already is.
